[Issue-0000] DataState.Loading refactor
[Summary]
Loading state should display both the loading indicator, and whatever UI
is already showing. This means remembering the last non-Loading state.
Got around this issue by using remember { } in Composables, but this
issue will come up again in Swift UI, as all UI are emitted as distinct,
whereas the Loading UI is a combination of the Loading indicator and the
current UI state. We got away with this fine with the Android View
system and UI Kit, as they implicitly held state. But we can no longer,
and should no longer, rely on Views to hold state.

[Fix]
Make DataState.Loading hold on to the last non-Loading state emitted, so
they can both be displayed at the same time. This could also be solved
via a SharedFlow with a cache of 2. However, I didn't want to worry
about translating that cache to Combine or RxSwift if we adopt that
later on. Adding a single parameter to Loading seemed like the simplest
solution.

- dependency update: updated AGP to 7.1.0-alpha03 to keep current with
  Android Studio Bumblebee 2021.1.1 Canary 3

[Testing]
- `./gradlew :app:build`
- `./gradlew :shared:build`
- `xcodebuild -workspace ios/KaMPKitiOS.xcworkspace -scheme KaMPKitiOS
    -sdk iphoneos -configuration Debug build -destination name="iPhone 8"`
- manual testing of clicking and swipe-to-refresh on iOS and Android